requirements.txt
pandas
numpy
scikit-learn
dask[complete]
dask-ml
pytest
pytest-asyncio
pytest-cov
pytest-mock
hypothesis
sqlalchemy
pyyaml
tqdm
beautifulsoup4
requests
boto3
spacy
gensim
matplotlib
seaborn
plotly
nltk
imbalanced-learn
cerberus
textblob
scipy
